+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add support for marking scripts as JavaScript modules
 * Add support for assigning element ID and/or class
 * Add support for default site langauge, and page specific languages
+* Bug fix to remove calls to deprecated Permission model
 
 ## 3.6.0